+SessionHelper
25 Выделяем специализированные классыпомощники.mp4
