+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/appmanager/ApplicationManager.java
@@ -2,7 +2,6 @@ package ru.stqa.qajava.addressbook.appmanager;
 
 import org.openqa.selenium.*;
 import org.openqa.selenium.chrome.ChromeDriver;
-import ru.stqa.qajava.addressbook.model.GroupData;
 
 import java.util.concurrent.TimeUnit;
 
@@ -10,6 +9,10 @@ import static org.testng.Assert.fail;
 
 public class ApplicationManager {
   public WebDriver driver;
+
+  private SessionHelper sessionHelper;
+  private NavigationHelper navigationHelper;
+  private GroupHelper groupHelper;
   public String baseUrl;
   public boolean acceptNextAlert = true;
   public StringBuffer verificationErrors = new StringBuffer();
@@ -20,46 +23,12 @@ public class ApplicationManager {
     baseUrl = "https://www.katalon.com/";
     driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
     driver.get("http://localhost/addressbook/group.php");
-    login("admin", "secret");
+    groupHelper = new GroupHelper(driver);
+    navigationHelper = new NavigationHelper(driver);
+    sessionHelper = new SessionHelper(driver);
+    sessionHelper.login("admin", "secret");
   }
 
-  private void login(String username, String passord) {
-    driver.findElement(By.name("user")).click();
-    driver.findElement(By.name("user")).clear();
-    driver.findElement(By.name("user")).sendKeys(username);
-    driver.findElement(By.name("pass")).click();
-    driver.findElement(By.name("pass")).clear();
-    driver.findElement(By.name("pass")).sendKeys(passord);
-    driver.findElement(By.xpath("(.//*[normalize-space(text()) and normalize-space(.)='Password:'])[1]/following::input[2]")).click();
-  }
-
-  public void reternToGroupPage() {
-    driver.findElement(By.linkText("group page")).click();
-  }
-
-  public void submitGroupCreation() {
-    driver.findElement(By.name("submit")).click();
-  }
-
-  public void fillGroupForm(GroupData groupData) {
-    driver.findElement(By.name("group_name")).click();
-    driver.findElement(By.name("group_name")).clear();
-    driver.findElement(By.name("group_name")).sendKeys(groupData.getGroupName());
-    driver.findElement(By.name("group_header")).click();
-    driver.findElement(By.name("group_header")).clear();
-    driver.findElement(By.name("group_header")).sendKeys(groupData.getGroupHeader());
-    driver.findElement(By.name("group_footer")).click();
-    driver.findElement(By.name("group_footer")).clear();
-    driver.findElement(By.name("group_footer")).sendKeys(groupData.getGroupFooter());
-  }
-
-  public void initGroupCreation() {
-    driver.findElement(By.name("new")).click();
-  }
-
-  public void gotoGroupPage() {
-    driver.findElement(By.linkText("groups")).click();
-  }
 
   public void stop() {
     driver.quit();
@@ -102,11 +71,11 @@ public class ApplicationManager {
     }
   }
 
-  public void deleteSelectedGroups() {
-    driver.findElement(By.name("delete")).click();
+  public GroupHelper getGroupHelper() {
+    return groupHelper;
   }
 
-  public void selectGroup() {
-    driver.findElement(By.xpath("(.//*[normalize-space(text()) and normalize-space(.)='test1'])[1]/input[1]")).click();
+  public NavigationHelper getNavigationHelper() {
+    return navigationHelper;
   }
 }
