+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/appmanager/GroupHelper.java
@@ -1,0 +1,45 @@
+package ru.stqa.qajava.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import ru.stqa.qajava.addressbook.model.GroupData;
+
+public class GroupHelper {
+  private WebDriver driver;
+
+  public GroupHelper(WebDriver driver) {
+    this.driver=driver;
+  }
+
+  public void reternToGroupPage() {
+    driver.findElement(By.linkText("group page")).click();
+  }
+
+  public void submitGroupCreation() {
+    driver.findElement(By.name("submit")).click();
+  }
+
+  public void fillGroupForm(GroupData groupData) {
+    driver.findElement(By.name("group_name")).click();
+    driver.findElement(By.name("group_name")).clear();
+    driver.findElement(By.name("group_name")).sendKeys(groupData.getGroupName());
+    driver.findElement(By.name("group_header")).click();
+    driver.findElement(By.name("group_header")).clear();
+    driver.findElement(By.name("group_header")).sendKeys(groupData.getGroupHeader());
+    driver.findElement(By.name("group_footer")).click();
+    driver.findElement(By.name("group_footer")).clear();
+    driver.findElement(By.name("group_footer")).sendKeys(groupData.getGroupFooter());
+  }
+
+  public void initGroupCreation() {
+    driver.findElement(By.name("new")).click();
+  }
+
+  public void deleteSelectedGroups() {
+    driver.findElement(By.name("delete")).click();
+  }
+
+  public void selectGroup() {
+    driver.findElement(By.xpath("(.//*[normalize-space(text()) and normalize-space(.)='test1'])[1]/input[1]")).click();
+  }
+}
