+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/appmanager/NavigationHelper.java
@@ -1,0 +1,16 @@
+package ru.stqa.qajava.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class NavigationHelper {
+  private WebDriver driver;
+
+  public NavigationHelper(WebDriver driver) {
+    this.driver=driver;
+  }
+
+  public void gotoGroupPage() {
+    driver.findElement(By.linkText("groups")).click();
+  }
+}
